add tag field to anonymous enums and bump up default tag size to u32 for enums
Great work - now I want you to do two relatively simple tasks related to enum tags.
Basically, I want you to:
1. Bump up the default enum tag size from `u16` to `u32`. Though I find it hard to believe that there's any case where an enum would need to have over 65 thousand variants, a `u32` gives us more breathing room AND gives us better odds of CPU optimization. The cost is pretty much negligible.
2. Add a `tag` field to anonymous enums. Just like regular enums, anonymous enums also have tags. They should be explicitly accessible by the developer through a `tag` field. Same logic as regular enums. This is likely the only field that anonymous enums are going to have, because anything more than that requires discriminating the variants. For the time being, we do not allow the `tag` field size to be different than whatever the compiler chooses (unlike regular enums, where you can override the compiler preference through the enum header).
I think this is a relatively simple change, so no plan required. Also, update documentation related to these changes. Make sure to read `ARCHITECTURE.md` so you don't explore the entire project for no good reason.

+++ compiler/omega-analyzer/src/error/mod.rs
@@ -323,7 +323,7 @@ impl fmt::Display for TypeResolutionError {
             ),
             Self::AnonymousEnumTooManyMembers { count } => write!(
                 f,
-                "an anonymous enum has {count} distinct members, but its tag is a 'u16' and can \
+                "an anonymous enum has {count} distinct members, but its tag is a 'u32' and can \
                  only distinguish {} of them",
                 crate::resolved_type::ResolvedAnonymousEnum::MAX_MEMBERS
             ),
